+++ src/main/java/me/jddev0/ep/block/entity/base/WorkerMachineBlockEntity.java
@@ -45,7 +45,7 @@ public abstract class WorkerMachineBlockEntity<W>
         return new EnergizedPowerEnergyStorage(baseEnergyCapacity) {
             @Override
             public long getCapacityAsLong() {
-                return Math.max(1, (int)Math.ceil(capacity * upgradeModuleInventory.getModifierEffectProduct(
+                return Math.max(1, (long)Math.ceil(capacity * upgradeModuleInventory.getModifierEffectProduct(
                         UpgradeModuleModifier.ENERGY_CAPACITY)));
             }
